+++ KeyboardKit/ObjC/KeyInputStrings.m
@@ -1,0 +1,14 @@
+// Douglas Hill, August 2024
+
+#import "KeyInputStrings.h"
+
+NSString *_KBDKeyInputUpArrow(void) { return UIKeyInputUpArrow; }
+NSString *_KBDKeyInputDownArrow(void) { return UIKeyInputDownArrow; }
+NSString *_KBDKeyInputLeftArrow(void) { return UIKeyInputLeftArrow; }
+NSString *_KBDKeyInputRightArrow(void) { return UIKeyInputRightArrow; }
+NSString *_KBDKeyInputEscape(void) { return UIKeyInputEscape; }
+NSString *_KBDKeyInputPageUp(void) { return UIKeyInputPageUp; }
+NSString *_KBDKeyInputPageDown(void) { return UIKeyInputPageDown; }
+NSString *_KBDKeyInputHome(void) { return UIKeyInputHome; }
+NSString *_KBDKeyInputEnd(void) { return UIKeyInputEnd; }
+NSString *_KBDKeyInputDelete(void) { return UIKeyInputDelete; }