+++ KeyboardKitDemo/FirstResponderManagement.swift
@@ -7,12 +7,10 @@ import KeyboardKit
 let firstResponderDidChangeNotification = Notification.Name("KBDFirstResponderDidChange")
 
 /// An abstract view controller that updates first responder when it appears.
-class FirstResponderViewController: UIViewController {
+class FirstResponderViewController: InitialiserClearingViewController {
     init() {
-        super.init(nibName: nil, bundle: nil)
+        super.init(onMainActor: ())
     }
-
-    @available(*, unavailable) required init?(coder: NSCoder) { preconditionFailure() }
 
     var windowIWasIn: UIWindow?