+++ KeyboardKitDemo/TripleColumnSplitViewController.swift
@@ -8,7 +8,7 @@ import KeyboardKit
 /// This class also demonstrates one possible way to show the focused column visually to the user:
 /// by showing the title in the navigation controller for the focused column darker than the titles
 /// in the other columns.
-class TripleColumnSplitViewController: UIViewController, TripleColumnListViewControllerDelegate, KeyboardSplitViewControllerDelegate, DismissModalActionPerformer {
+class TripleColumnSplitViewController: InitialiserClearingViewController, TripleColumnListViewControllerDelegate, KeyboardSplitViewControllerDelegate, DismissModalActionPerformer {
     private let innerSplitViewController: KeyboardSplitViewController
 
     // The primary and supplementary would ideally use the sidebar and sidebarPlain styles.
@@ -29,7 +29,6 @@ class TripleColumnSplitViewController: UIViewController, TripleColumnListViewCon
     private let secondaryNavigationController: KeyboardNavigationController
 
     @available(*, unavailable) override var splitViewController: UISplitViewController? { nil }
-    @available(*, unavailable) required init?(coder: NSCoder) { preconditionFailure() }
 
     init() {
         innerSplitViewController = KeyboardSplitViewController(style: .tripleColumn)
@@ -42,7 +41,7 @@ class TripleColumnSplitViewController: UIViewController, TripleColumnListViewCon
         innerSplitViewController.setViewController(supplementaryNavigationController, for: .supplementary)
         innerSplitViewController.setViewController(secondaryNavigationController, for: .secondary)
 
-        super.init(nibName: nil, bundle: nil)
+        super.init(onMainActor: ())
 
         innerSplitViewController.delegate = self
         primaryList.delegate = self